apps/backend/requirements.txt
python-dotenv
python-dateutil
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-Limiter
Flask-WTF
Flask-JWT-Extended
authlib
unidecode
gunicorn
psycopg2
google-cloud-storage
google-cloud-scheduler
google-cloud-pubsub
cryptography
pywebpush